attached_assets/BrainScanAI/pyproject.toml

[tool.poetry]
name = "repl_nix_brain-tumor-detector"
version = "0.1.0"
description = "CNN-Based Brain Tumor Detection System"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.26.0"
numpy = "^1.26.0"
pandas = "^2.1.0"
pillow = "^10.0.0"
opencv-python-headless = "^4.8.0"
plotly = "^5.17.0"
fpdf = "^1.7.2"
tensorflow = "^2.15.0"
scikit-learn = "^1.3.0"
matplotlib = "^3.8.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
